debugging

+++ process_table_shm.h
@@ -3,7 +3,7 @@
 
 typedef struct
 {
-   long pid;
+   int pid;
    int priority;
    int cpu_time;
    int total_sec;
